Rename class back to Duke and clean up style
