+++ build.gradle
@@ -18,7 +18,7 @@ repositories {
 
 application {
     // Change this to your main class.
-    mainClassName = "duke.Akshay"
+    mainClassName = "duke.Duke"
 }
 
 run {
