+++ data/duke.txt
@@ -1,2 +1,4 @@
 D , 1 , asdl , 2019-10-09
 D , 0 , 91239 , 2019-10-10
+T , 1 , read book
+E , 0 , collect book  , 2019-03-12
