+++ src/test/java/duke/TodoTest.java
@@ -2,12 +2,12 @@ package duke;
 
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class TodoTest {
 
     @Test
     void testToString() {
-        assertEquals("[T][\u2713] Blah", new Todo("Blah", true).toString());
+        assertEquals("[T][✓] Blah", new Todo("Blah", true).toString());
     }
 }
